Split Windows and Unix implementations into two files
